renamed porxy_map -> mapping
