+++ autoforce/core/parameters/__init__.py
@@ -1,5 +1,6 @@
 # +
 from .chempar import ChemPar, ReducedPar
 from .cutoff import Cutoff
+from .mapping import AsymmetricMapping, SymmetricMapping
 
-__all__ = ["ChemPar", "ReducedPar", "Cutoff"]
+__all__ = ["ChemPar", "ReducedPar", "Cutoff", "SymmetricMapping", "AsymmetricMapping"]
